pyproject.toml
[project]
name = "autoregressive-nano-tabpfn"
version = "0.1.0"
description = "Efficient TabPFN with flex_attention training and Triton inference"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.5.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "h5py>=3.8.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "seaborn>=0.12.0",
    "pytest>=7.0.0",
]
# Triton is only available on Linux with CUDA
cuda = [
    "triton>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autoregressive_nano_tabpfn"]

[tool.uv]
dev-dependencies = [
    "seaborn>=0.12.0",
    "pytest>=7.0.0",
]
